杨毅晨:更新杨毅晨组YangYiChen项目,更新时间:2018-12-23 23:26!!!

+++ YangYiChen/src/main/java/com/cxb/yyc/controller/TeacherSetTestController.java
@@ -5,34 +5,44 @@ import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 
 import com.cxb.yyc.entity.TeacherCreateTest;
+import com.cxb.yyc.service.TeacherSetChoiceQuestionService;
+import com.cxb.yyc.service.TeacherSetGapfillingQuestionService;
 import com.cxb.yyc.service.TeacherSetTestService;
-@CrossOrigin
+import com.cxb.yyc.service.TeacherSetTfngQuestionService;
+
 @RestController
 @RequestMapping(value="/teacherTest",name="发起测试")
+@CrossOrigin
 public class TeacherSetTestController {
 	/**
 	 * 教师测试业务实现类接口
 	 */
 	@Autowired
 	private TeacherSetTestService testService;
+	
+	@Autowired
+	private TeacherSetChoiceQuestionService teacherSetChoiceQuestionService;
+	
+	@Autowired
+	private TeacherSetTfngQuestionService teacherSetTfngQuestionService;
+	
+	@Autowired
+	private TeacherSetGapfillingQuestionService teacherSetGapfillingQuestionService;
 	/**
 	 * 教师发起测试
 	 * http://localhost:3060/YangYiChen/teacherTest/insertTeacherCreateTest?teachercratetestChoicenum=2&teachercratetestTestname=第一次章节测试 &teachercreatetestTesttime=30min&teachercreatetestEndtime=2019-01-12 00:00:00&teachercreatetestNumber=3&teachercratetestRandom=是&teachercreatetestExplain=测试学生对之前内容的掌握程度&teachercreatetestChapterId=1&teachercreatetestTeacherId=1
 	 * @param teacherCreateTest
 	 * @return
 	 */
-	@PostMapping(value="/insertTeacherCreateTest",name="教师发起测试")
+	@RequestMapping(value="/insertTeacherCreateTest",name="教师发起测试",method=RequestMethod.POST)
 	public Object insertTeacherCreateTest(@RequestBody TeacherCreateTest teacherCreateTest) {
 		Map<String, Object> map = new HashMap<String, Object>();
 		teacherCreateTest.setTeachercratetestState(0);
@@ -52,7 +62,7 @@ public class TeacherSetTestController {
 	 * @param teacherId
 	 * @return
 	 */
-	@GetMapping(value="/queryTeacherCreateTest",name="教师查询测试")
+	@RequestMapping(value="/queryTeacherCreateTest",name="教师查询测试",method=RequestMethod.GET)
 	public Object queryTeacherCreateTest(Integer teacherId) {
 		return testService.queryTeacherCreateTest(teacherId);
 	}
@@ -65,7 +75,7 @@ public class TeacherSetTestController {
 	 * @param state
 	 * @return
 	 */
-	@PutMapping(value="/updateTeacherCreateTestState",name="发布测试")
+	@RequestMapping(value="/updateTeacherCreateTestState",name="发布测试",method=RequestMethod.PUT)
 	public Object updateTeacherCreateTestState(@RequestParam(value="testId",required=false)Integer testId,@RequestParam(value="state",required=false)Integer state) {
 			Map<String, Object> map=new HashMap<String, Object>();
 			if (state==0) {
@@ -83,14 +93,14 @@ public class TeacherSetTestController {
 			return map;
 	}
 	/**
-	 * 根据教师发布测试主键发布测试
+	 * 根据教师发布测试主键取消测试
 	 * http://localhost:3060/YangYiChen/teacherTest/updateTeacherCreateTestState?testId=24&state=1
 	 * http://localhost:3011/yangyichen/YangYiChen/teacherTest/updateTeacherCreateTestStateForZero?testId=1&state=1
 	 * @param testId
 	 * @param state
 	 * @return
 	 */
-	@PutMapping(value="/updateTeacherCreateTestStateForZero",name="取消发布")
+	@RequestMapping(value="/updateTeacherCreateTestStateForZero",name="取消发布",method=RequestMethod.PUT)
 	public Object updateTeacherCreateTestStateForZero(@RequestParam(value="testId",required=false)Integer testId,@RequestParam(value="state",required=false)Integer state) {
 			Map<String, Object> map=new HashMap<String, Object>();
 			if (state==1) {
@@ -114,15 +124,24 @@ public class TeacherSetTestController {
 	 * @param testId
 	 * @return
 	 */
-	@DeleteMapping(value="/deleteTest",name="删除自己创建的测试题")
+	@RequestMapping(value="/deleteTest",name="删除自己创建的测试题",method=RequestMethod.DELETE)
 	public Object deleteTest(@RequestParam(value="testId",required=false) Integer testId) {
 		Map<String, Object> map=new HashMap<String, Object>();
-		if (testService.deleteTest(testId)>0) {
-			map.put("success", true);
-			map.put("msg", "删除成功！");
+		int chapterId=testService.selectByTestIdForChapterId(testId);
+		int n1=teacherSetChoiceQuestionService.selectCountByChoiceChapterId(chapterId);
+		int n2=teacherSetTfngQuestionService.selectCountByTfngChapterId(chapterId);
+		int n3=teacherSetGapfillingQuestionService.selectCountByGapfillingChapterId(chapterId);
+		if (n1==0&&n2==0&&n3==0) {
+			if (testService.deleteTest(testId)>0) {
+				map.put("success", true);
+				map.put("msg", "删除成功！");
+			}else {
+				map.put("success", false);
+				map.put("msg", "删除失败！");
+			}
 		}else {
 			map.put("success", false);
-			map.put("msg", "删除失败！");
+			map.put("msg", "该测试题下还有题型未删除完，请先删除所有题型！");
 		}
 		return map;
